separated database functions into modular files
