+++ database/seedDB.js
@@ -1,22 +1,7 @@
 const mongoose = require('mongoose');
-const { genAllData } = require('./dataGen');
+const { genAllData, Recommendations } = require('./dataGen');
 
 mongoose.connect('mongodb://localhost/recommendations');
-
-const recSchema = new mongoose.Schema({
-  title: String,
-  description: String,
-  price: String,
-  photo_url: String,
-});
-
-const listingSchema = new mongoose.Schema({
-  listing_id: { type: Number, unique: true },
-  listing_title: String,
-  recommendations: [recSchema],
-});
-
-const Recommendations = mongoose.model('recommendation', listingSchema);
 
 const seedDB = (data) => {
   const promise = Recommendations.create(data);
@@ -25,27 +10,8 @@ const seedDB = (data) => {
   });
 };
 
-const find = (listing, callback) => {
-  Recommendations.find((err, recommendation) => {
-    if (err) {
-      console.log(err);
-    } else {
-      callback(recommendation);
-    }
-  }).where('listing_id').equals(listing);
-};
 
-const dropDB = () => {
-  mongoose.connection.collection.recommendations.drop((err) => {
-    if (err) {
-      console.log(err);
-    }
-    console.log('collection dropped');
-  });
-};
+seedDB(genAllData(5));
 
-// dropDB();
 exports.seedDB = seedDB;
-exports.find = find;
-exports.Recommendations = Recommendations;
-exports.dropDB = dropDB;
+
